Rename elements/browserElements/api.js -> utils/text/SkriptShaper.js
