+++ test/text/index.html
@@ -58,7 +58,7 @@
         }
 
     </style>
-    <script src="../../player/js/elements/browserElements/api.js"></script>
+    <script src="../../player/js/utils/text/SkriptShaper.js"></script>
 </head>
 <body style="margin: 0px; padding: 0px; border: 0px; font-family:'Noto Sans', 'Noto Sans Arabic', 'Noto Sans Emoji', 'Noto Sans Bengali'; font-size: 20px;">
     <div class="row">
